refactored index
