+++ apps/server/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "display": "Server",
   "compilerOptions": {
+    "resolveJsonModule": true,
     "lib": ["DOM", "DOM.Iterable", "ES2019"],
     "composite": false,
     "declaration": false,
