--- apps/web/components/ui/MediaCard/types.ts
@@ -1,5 +1,0 @@
-export type MediaItem = {
-  text: string;
-  image: string;
-  alignment: string;
-};
