+++ apps/web/composables/useHomepageData/homepageTemplateData.json
@@ -1,0 +1,74 @@
+{
+  "id": 100,
+  "hero": [
+    {
+      "image": {
+        "lg": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/homepage-hero-headphones.avif",
+        "md": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-md.avif",
+        "sm": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-sm.avif",
+        "xs": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-xs.avif",
+        "alt": "Headphones"
+      },
+      "tagline": "Feel the music",
+      "taglineColor": "#000000",
+      "heading": "Your Sound, Elevated",
+      "headingColor": "#000000",
+      "description": "Immerse yourself in rich, crystal-clear audio with our cutting-edge headphones. Designed for the ultimate listening experience, whether you're a casual listener or an audiophile. Discover the perfect blend of style, comfort, and sound quality that elevates your music to new heights.\n\n",
+      "descriptionColor": "#000000",
+      "callToAction": "Order Now",
+      "link": ""
+    },
+    {
+      "image": {
+        "lg": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/homepage-hero-headphones.avif",
+        "md": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-md.avif",
+        "sm": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-sm.avif",
+        "xs": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-xs.avif",
+        "alt": "Headphones"
+      },
+      "tagline": "Experience Sound Freedom",
+      "taglineColor": "#000000",
+      "heading": "Wireless. Effortless. Seamless.",
+      "headingColor": "#000000",
+      "description": "Unleash your audio with our state-of-the-art wireless earbuds. Designed for all-day comfort and uncompromised sound quality, these earbuds deliver crisp highs and deep bass, letting you enjoy your music without any distractions. Discover freedom with a perfect fit, long battery life, and intuitive controls.",
+      "descriptionColor": "#000000",
+      "callToAction": "Shop Earbuds",
+      "link": ""
+    },
+    {
+      "image": {
+        "lg": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/homepage-hero-headphones.avif",
+        "md": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-md.avif",
+        "sm": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-sm.avif",
+        "xs": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-xs.avif",
+        "alt": "Headphones"
+      },
+      "tagline": "Amplify Your Space",
+      "taglineColor": "#000000",
+      "heading": "Big Sound, Compact Design",
+      "headingColor": "#000000",
+      "description": "Transform your space with our portable speakers that pack a punch. Crafted for superior sound performance, these speakers are perfect for home or on the go. With easy connectivity and a sleek design, elevate your listening experience whether you're indoors or outdoors.",
+      "descriptionColor": "#000000",
+      "callToAction": "Browse Speakers",
+      "link": ""
+    }
+  ],
+  "valueProposition": [
+    {
+      "text": "<div class='flex flex-col mt-5 sm:mt-20 mt-0 sm:p-0 p-5 text-center sm:text-left'><span class='text-xl font-bold mb-2'>Experience the Future of Sound</span><h2 class='text-2xl font-semibold mb-4'>Redefine Your Listening Experience</h2><p class='typography-text-sm md:typography-text-lg mb-6 padding-right-desktop'>Our latest collection of headphones is designed to deliver unparalleled audio precision, with deep bass, clear highs, and an immersive experience for every genre of music. Combining sleek design, comfort, and cutting-edge technology, these headphones are made for those who refuse to compromise on sound quality.</p><ul class='list-disc list-inside typography-text-sm md:typography-text-lg '><li>Premium, studio-quality sound</li><li>Comfortable fit for extended listening</li><li>Long-lasting battery life</li><li>Seamless wireless connectivity</li></ul></div>",
+      "image": "https://cdn02.plentymarkets.com/mevofvd5omld/frontend/headphones-mediacard.avif",
+      "alt": "Headphones",
+      "alignment": "left"
+    }
+  ],
+  "featured": [
+    {
+      "headline": "",
+      "categoryId": 1
+    },
+    {
+      "headline": "",
+      "categoryId": 2
+    }
+  ]
+}
