+++ apps/web/composables/useHomepageData/useHomepageData.ts
@@ -1,0 +1,81 @@
+import { ref, watch, computed } from 'vue';
+import { HeroItem, SizeKey } from '~/composables/useHomepageData/types';
+import { MediaItem } from '~/composables/useHomepageData/types';
+import { useCategoryTree, useCategoryTemplate } from '~/composables';
+import homepageTemplateData from './homepageTemplateData.json';
+
+const resolveImage = (imageSizes: Record<SizeKey, string>, sizeKey: SizeKey): string => {
+  return imageSizes[sizeKey] || '';
+};
+
+export default async function useHomepageData() {
+  const viewport = useViewport();
+  const { data: categoryTree } = useCategoryTree();
+  const recommendedProductsCategoryId = ref('');
+
+  const runtimeConfig = useRuntimeConfig();
+  const homepageTemplate = ref(homepageTemplateData);
+
+  const homepageCategoryId = runtimeConfig.public.homepageCategoryId;
+  const { fetchCategoryTemplate } = useCategoryTemplate();
+
+  if (typeof homepageCategoryId === 'number') {
+    const { data } = await fetchCategoryTemplate(homepageCategoryId);
+    const parsedData = JSON.parse(data);
+    if (parsedData) {
+      homepageTemplate.value = {
+        id: parsedData.id,
+        hero: parsedData.hero || [],
+        valueProposition: parsedData.valueProposition,
+        featured: parsedData.featured,
+      };
+    }
+  }
+
+  const getCurrentSizeKey = (): SizeKey => viewport.breakpoint.value as SizeKey;
+
+  const mediaData = computed(() =>
+    homepageTemplate.value.valueProposition.map((media: MediaItem) => ({
+      image: media.image,
+      text: media.text,
+      alignment: media.alignment,
+      alt: media.alt,
+    })),
+  );
+
+  const formattedHeroItems = computed<HeroItem[]>(() => {
+    const currentSizeKey = getCurrentSizeKey();
+    return homepageTemplate.value.hero.map((item) => {
+      return {
+        image: resolveImage(item.image, currentSizeKey),
+        tagline: item.tagline || '',
+        heading: item.heading || '',
+        description: item.description || '',
+        callToAction: item.callToAction || '',
+        link: item.link || '',
+        backgroundSizes: {
+          lg: { width: '4000', height: '600' },
+          md: { width: '1024', height: '600' },
+          sm: { width: '640', height: '752' },
+          xs: { width: '250', height: '250' },
+        },
+        actualBackgroundSize: currentSizeKey,
+      } as HeroItem;
+    });
+  });
+
+  watch(
+    () => categoryTree.value,
+    () => {
+      const firstCategoryId = categoryTree.value?.[0]?.id;
+      if (firstCategoryId) recommendedProductsCategoryId.value = firstCategoryId.toString();
+    },
+    { immediate: true },
+  );
+
+  return {
+    formattedHeroItems,
+    mediaData,
+    recommendedProductsCategoryId,
+  };
+}
